добавил тест для тестирования все страниц админки на ошибки

+++ protected/modules/banners/models/Banner.php
@@ -8,6 +8,10 @@ class Banner extends ActiveRecordModel
 
     const PAGE_SIZE = 10;
 
+    public $date_active;
+
+    public $src;
+
 
 	public static function model($className=__CLASS__)
 	{
@@ -31,7 +35,7 @@ class Banner extends ActiveRecordModel
 	{
 		return array(
 			array('name', 'required'),
-			array('is_active', 'numerical', 'integerOnly' => true),
+			array('is_active, page_id', 'numerical', 'integerOnly' => true),
 			array('name', 'length', 'max' => 50),
 			array(
                 'image',
@@ -39,20 +43,63 @@ class Banner extends ActiveRecordModel
                 'types' => 'jpg, jpeg, png, gif',
                 'allowEmpty' => true
             ),
+            array('date_start', 'validateDateInterInterVal'),
 			array('url', 'length', 'max' => 500),
-			array('date_start, date_end', 'safe'), array('roles', 'safe', 'on' => 'cre'),
+			array('date_start, date_end, date_active', 'safe'),
+            array('roles', 'safe'),
 			array('id, name, image, url, is_active, date_start, date_end', 'safe', 'on' => 'search'),
 		);
 	}
+
+
+    public function validateDateInterInterVal($attr)
+    {
+        if ($this->date_active)
+        {
+            if (!$this->$attr || !$this->date_end)
+            {
+                $this->addError($attr, 'Укажите и дату начала и дату окончания показа!');
+                return;
+            }
+
+            if ($this->$attr && $this->date_end)
+            {   //v($this->$attr);
+                //v($this->date_end);
+                //die;
+                $date_start = strtotime($this->$attr);
+                $date_end   = strtotime($this->date_end);
+
+                if ($date_start > $date_end)
+                {
+                    $this->addError($attr, 'Дата окончания показа не должна быть раньше даты начала показа!');
+                }
+            }
+        }
+    }
 
 
 	public function relations()
 	{
 		return array(
 			'banners_roles' => array(self::HAS_MANY, 'BannerRole', 'banner_id'),
-            'roles' => array(self::HAS_MANY, 'AuthItem', 'role', 'through' => 'banners_roles')
+            'roles' => array(self::HAS_MANY, 'AuthItem', 'role', 'through' => 'banners_roles'),
+            'page'  => array(self::BELONGS_TO, 'Page', 'page_id')
 		);
 	}
+
+
+    public function behaviors()
+    {
+        $behaviors = parent::behaviors();
+        $behaviors['sortable'] = array(
+            'class'=> 'application.components.activeRecordBehaviors.SortableBehavior'
+        );
+        $behaviors['MaxMinBehavior'] = array(
+            'class' => 'application.components.activeRecordBehaviors.MaxMinBehavior'
+        );
+
+        return $behaviors;
+    }
 
 
 	public function search()
@@ -65,6 +112,7 @@ class Banner extends ActiveRecordModel
 		$criteria->compare('is_active', $this->is_active);
 		$criteria->compare('date_start', $this->date_start, true);
 		$criteria->compare('date_end', $this->date_end, true);
+        $criteria->order = 'priority';
 
 		return new ActiveDataProvider(get_class($this), array(
 			'criteria' => $criteria
@@ -104,17 +152,55 @@ class Banner extends ActiveRecordModel
 
     public function afterSave()
     {
-        BannerRole::model()->deleteAll('banner_id = ' . $this->id);
-
-        if ($this->roles)
+        if (in_array($this->scenario, array(self::SCENARIO_CREATE, self::SCENARIO_UPDATE)))
         {
-            foreach ($this->roles as $role)
-            {  
-                $banner_role = new BannerRole;
-                $banner_role->banner_id = $this->id;
-                $banner_role->role = $role;
-                $banner_role->save();
+            BannerRole::model()->deleteAll('banner_id = ' . $this->id);
+
+            if ($this->roles)
+            {
+                foreach ($this->roles as $role)
+                {
+                    $banner_role = new BannerRole;
+                    $banner_role->banner_id = $this->id;
+                    $banner_role->role = $role;
+                    $banner_role->save();
+                }
             }
         }
+
+        parent::afterSave();
+    }
+
+
+    public function attributeLabels()
+    {
+        $labels = parent::attributeLabels();
+        $labels['roles'] = 'Отображается для ролей';
+        return $labels;
+    }
+
+
+    public function getHref()
+    {
+        if ($this->page)
+        {
+            return $this->page->href;
+        }
+        else
+        {
+            return $this->url;
+        }
+    }
+
+
+
+    public function beforeSave()
+    {
+        if ($this->isNewRecord)
+        {
+            $this->setDefaultPriority();
+        }
+
+        return parent::beforeSave();
     }
 }
